requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
httpx==0.26.0
python-dotenv==1.0.0
tiktoken==0.5.2
boto3==1.34.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-auth
email-validator
